requirements.txt
mysql-connector-python
pymongo
bcrypt
python-dotenv
tk
